rename to DefaultWorkspace
