+++ src/core/trulens/core/workspace/__init__.py
@@ -1,4 +1,4 @@
 from trulens.core.workspace.base import BaseWorkspace
-from trulens.core.workspace.sqlalchemy import SQLAlchemyWorkspace
+from trulens.core.workspace.default import DefaultWorkspace
 
-__all__ = ["BaseWorkspace", "SQLAlchemyWorkspace"]
+__all__ = ["BaseWorkspace", "DefaultWorkspace"]
